fix when exponent is absent in optional decimal
According to FAST specification (sec 10.5.1) :

The presence of the mantissa field and any related
bits in the presence map are dependent on the
presence of the exponent. The mantissa field appears
in the stream iff the exponent value is considered
present. If the mantissa has an operator that
requires a bit in the presence map, this bit is
present iff the exponent value is considered present.